feat(client): qna 바로가기 bar 생성
