+++ apps/client/src/components/qna/QnaPostCard/index.tsx
