--- apps/client/src/constants/qna-category.ts
@@ -1,22 +1,0 @@
-export const QNA_CATEGORY = [
-  {
-    id: 0,
-    title: '전체',
-    emoji: '🖥️',
-  },
-  {
-    id: 1,
-    title: '개발',
-    emoji: '💻',
-  },
-  {
-    id: 2,
-    title: '취업',
-    emoji: '👔',
-  },
-  {
-    id: 3,
-    title: '고민',
-    emoji: '🤔',
-  },
-];
