--- apps/client/src/constants/qna-sort.ts
@@ -1,14 +1,0 @@
-export const QNA_SORT = [
-  {
-    id: 0,
-    title: '최신순',
-  },
-  {
-    id: 1,
-    title: '공감순',
-  },
-  {
-    id: 2,
-    title: '조회순',
-  },
-];
