+++ apps/client/src/constants/qna.ts
@@ -64,3 +64,48 @@ export const QNA = [
       '프론트엔드 취준생인데요 지금 알고리즘 프로그래머스 기준으로 2단계까지는 풀 수 있는데 코딩 테스트를 볼 때 테스트 케이스 모두 통과하고 제출하는데 항상 탈락해요.. 코딩 테스트는 풀 수만 있으면 된다고 들었는데 아닌가요? 게다가 테스트 케이스 만점이 탈락이라면 대체 어떻게 해야 합격인걸까요? 합격의 기준이 뭔지 모르겠습니다.. 혹시 코드 성능이 합불에 영향을 미치고 있다면 코테 보는 중에는 직접 측정할 수도 없는데 이 부분은 어떻게 스스로 개선해야할까요?',
   },
 ];
+
+export const QNA_SORT = [
+  {
+    id: 0,
+    title: '최신순',
+  },
+  {
+    id: 1,
+    title: '공감순',
+  },
+  {
+    id: 2,
+    title: '조회순',
+  },
+];
+
+export const QNA_CATEGORY = [
+  {
+    id: 0,
+    title: '전체',
+    emoji: '🖥️',
+  },
+  {
+    id: 1,
+    title: '개발',
+    emoji: '💻',
+  },
+  {
+    id: 2,
+    title: '취업',
+    emoji: '👔',
+  },
+  {
+    id: 3,
+    title: '고민',
+    emoji: '🤔',
+  },
+];
+
+export const QNA_POSITION_LIST = [
+  { queryParams: 'all', name: '전체' },
+  { queryParams: 'develop', name: '개발' },
+  { queryParams: 'work', name: '취업' },
+  { queryParams: 'worry', name: '고민' },
+];
